Small refactoring after review
